moved functions from __init__.py to utils.py
